backend/requirements.txt
fastapi==0.95.2
uvicorn==0.27.1
sqlalchemy==2.0.27
asyncpg==0.29.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
bcrypt==4.0.1
pydantic==1.10.13
python-dotenv==1.0.1
aiofiles==23.2.1
alembic==1.13.1
sqladmin==0.16.0
jinja2==3.1.2
itsdangerous==2.1.2
email-validator==2.1.0.post1 